+++ src/types/property.ts
@@ -1,17 +1,6 @@
-export type IPropertyItem = {
-  id: string;
-  sale_rent: string | null;
-  saleRent: string | null;
-  status: string | null;
-  country: string;
-  city: string;
-  price: number;
-  list_selling_price_amount: number;
-  media: { file: { thumbnails: { medium: string } } }[];
-  photo: string;
-};
+import { QobrixProperty } from './qobrix';
 
-export type IPropertyList = IPropertyItem[];
+export type IPropertyList = QobrixProperty[];
 
 export type IProperties = {
   data: IPropertyList;
