+++ src/types/qobrix/property/qobrix-propety-type.type.ts
@@ -1,0 +1,11 @@
+type QobrixPropertyTypeResponse = {
+  data: QobrixPropertyType[];
+};
+
+type QobrixPropertyType = {
+  base_type: string;
+  code: string;
+  name: string;
+};
+
+export { type QobrixPropertyTypeResponse, type QobrixPropertyType };
