--- src/types/qobrix/qobrix-propety-type.type.ts
@@ -1,9 +1,0 @@
-type QobrixPropertyType = {
-  data: {
-    base_type: string;
-    code: string;
-    name: string;
-  }[];
-};
-
-export { type QobrixPropertyType };
